improved project architecture
