+++ Pr2503/Sources/Scenes/Controller/ViewController.swift
@@ -1,15 +1,14 @@
 import UIKit
 
 class ViewController: UIViewController {
+    
     @IBOutlet weak var button: UIButton!
     @IBOutlet weak var textField: UITextField!
     @IBOutlet weak var label: UILabel!
     @IBOutlet weak var myButton: UIButton!
     @IBOutlet weak var spinner: UIActivityIndicatorView!
     @IBOutlet weak var mySwitch: UISwitch!
-    
-    var len = Int()
-    
+        
     var isBlack: Bool = false {
         didSet {
             if isBlack {
@@ -27,7 +26,7 @@ class ViewController: UIViewController {
     }
     
     @IBAction func passGen(_ sender: Any) {
-        len = mySwitch.isOn ? 3 : 2
+        let len = mySwitch.isOn ? 3 : 2
         
         let pswdChars = String().printable
         let rndPswd = String((0..<len).compactMap{ _ in pswdChars.randomElement() })
@@ -41,9 +40,12 @@ class ViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
+        setupView()
+    }
+    
+    func setupView() {
         spinner.isHidden = true
-        
-        textField.isSecureTextEntry = true
+        textField.isSecureTextEntry = false
     }
     
     func bruteForce(passwordToUnlock: String) {
@@ -60,64 +62,13 @@ class ViewController: UIViewController {
         
         while password != passwordToUnlock {
             password = generateBruteForce(password, fromArray: ALLOWED_CHARACTERS)
-
-            print(password)
-            
         }
-        
-        print(password)
-        
+                
         DispatchQueue.main.async {
             self.spinner.stopAnimating()
             self.spinner.isHidden = true
             self.label.text = password
+            self.textField.isSecureTextEntry = true
         }
     }
 }
-
-
-
-extension String {
-    var digits:      String { return "0123456789" }
-    var lowercase:   String { return "abcdefghijklmnopqrstuvwxyz" }
-    var uppercase:   String { return "ABCDEFGHIJKLMNOPQRSTUVWXYZ" }
-    var punctuation: String { return "!\"#$%&'()*+,-./:;<=>?@[\\]^_`{|}~" }
-    var letters:     String { return lowercase + uppercase }
-    var printable:   String { return digits + letters + punctuation }
-
-
-
-    mutating func replace(at index: Int, with character: Character) {
-        var stringArray = Array(self)
-        stringArray[index] = character
-        self = String(stringArray)
-    }
-}
-
-func indexOf(character: Character, _ array: [String]) -> Int {
-    return array.firstIndex(of: String(character))!
-}
-
-func characterAt(index: Int, _ array: [String]) -> Character {
-    return index < array.count ? Character(array[index])
-                               : Character("")
-}
-
-func generateBruteForce(_ string: String, fromArray array: [String]) -> String {
-    var str: String = string
-
-    if str.count <= 0 {
-        str.append(characterAt(index: 0, array))
-    }
-    else {
-        str.replace(at: str.count - 1,
-                    with: characterAt(index: (indexOf(character: str.last!, array) + 1) % array.count, array))
-
-        if indexOf(character: str.last!, array) == 0 {
-            str = String(generateBruteForce(String(str.dropLast()), fromArray: array)) + String(str.last!)
-        }
-    }
-
-    return str
-}
-
